requirements.txt
opencv-python
requests
psycopg2-binary
python-dotenv
python-telegram-bot
mysql-connector-python
